# Speed up `std::fmt::Write::write_str` for `maud::Escaper`.