I'm seeing roughly 25% speedup on both the normal and complicated
benchmarks present in `benchmarks/benches/maud.rs` and
`benchmarks/benches/complicated_maud.rs` respectively.

Before:

    test render_template ... bench:             461 ns/iter (+/- 140)
    test render_complicated_template ... bench: 1,562 ns/iter (+/- 366)

After:

    test render_template ... bench:             350 ns/iter (+/- 69)
    test render_complicated_template ... bench: 1,145 ns/iter (+/- 633)

(The error margins are pretty high in the above snippets but I've
ran these benchmarks several times and the difference is consistent.)

<hr>

What do you think @lfairy?

I also tried using `str::slice_unchecked` (there is no easy way to do unchecked slicing for slices yet; see https://github.com/rust-lang/rust/issues/35729) but there was no measurable improvement.