@lfairy 2016-11-13
How does this compare when most of the characters need to be escaped? I know that some clever approaches fall over when given strings like `<<<<>>>><<<<>>>>` are involved.