@TheNeikos 2016-11-13
@lfairy Maybe that could be added to the benchmarks?