@utkarshkukreti 2016-11-14
@lfairy Good point. It's over 3.5x slower for escaping a long sequence of `<>` (13µs vs 3.5µs) and over 2x slower for escaping a long sequence of `<foo></foo>` (6.7µs vs 3µs). I guess the 2 benchmarks I compared in the commit message are the worst case for the current algorithm and best case for the new one. The question then is, what ratio of characters can we expect to require escaping in the real world and will this optimization be worth it for them?