@lfairy 2016-11-15
@TheNeikos I'd love to see a escaping benchmark! Feel free to file a PR for that.

@utkarshkukreti The values being escaped are often provided by untrusted users (e.g. comments on a blog), so we need to assume the worst case. For this reason I'm wary of approaches which vary so much based on the input. There are ways to narrow the gap (using instructions like PCMPESTRI) but that will add much more complexity.

For these reasons, I'm going to close this PR. Thanks for your contribution regardless!