@utkarshkukreti 2016-11-15
@lfairy Sounds reasonable. As for PCMPESTRI, I don't think it would be faster than the current implementation if the string contains only `<>&"` (but it would definitely be faster than the best case of this PR).