@utkarshkukreti 2016-11-15
With a tiny change I'm able to get a huge improvement for `<><><>...`: 14µs (this PR) -> 8.5µs (with diff below).

```diff
--- a/maud/src/lib.rs
+++ b/maud/src/lib.rs
@@ -181,8 +181,10 @@ impl<'a> fmt::Write for Escaper<'a> {
         for (i, &b) in bytes.iter().enumerate() {
             macro_rules! push {
                 ($str:expr) => {{
-                    unsafe {
-                        self.0.as_mut_vec().extend_from_slice(&bytes[last..i]);
+                    if i > last {
+                        unsafe {
+                            self.0.as_mut_vec().extend_from_slice(&bytes[last..i]);
+                        }
                     }
                     self.0.push_str($str);
                     last = i + 1;
```

The benchmark:

```diff
--- a/benchmarks/benches/maud.rs
+++ b/benchmarks/benches/maud.rs
@@ -39,3 +39,17 @@ fn render_template(b: &mut test::Bencher) {
         }
     });
 }
+
+#[bench]
+fn render_escape(b: &mut test::Bencher) {
+    // let text = "<foo></foo>";
+    let text = "<><><><><>";
+
+    b.iter(|| {
+        html! {
+            @for _ in 0..100 {
+                (text)
+            }
+        }
+    });
+}
```

(Posting here for future reference... if I ever find a way to make this reasonably fast in the worst case.)